requirements.txt
aiogram==3.25.0
aiosqlite==0.22.1
python-dotenv==1.1.0
